Revert "fix terminal resize and macOS zsh env"
